Add files via upload
This is the library, and source code needed to use the Si4703 fm radio module. The Arduino file labeled Si4703 is just preset for this specific chip, however this library covers the use of other IC's including the TEA5767. The sketch I modified uses a scan radio function which we need to add functions to buttons and a potentiometer accessible on the outside of the box.